+++ resources/big-config/package/tofu-backend/s3/backend.tf
@@ -1,0 +1,7 @@
+terraform {
+  backend "s3" {
+    bucket = "<{ s3-bucket }>"
+    key    = "<{ target-object }>.tfstate"
+    region = "<{ s3-region }>"
+  }
+}