+++ test/fixtures/target/tools-8/src/resources/io/github/amiorin/redis/tools/tofu/no-infra/main.tf
@@ -1,0 +1,9 @@
+output "params" {
+  value = {
+    ip = "<{ no-infra-compute-ip }>"
+    sudoer = "<{ no-infra-compute-sudoer }>"
+    uid = "<{ no-infra-compute-uid }>"
+    name = "<{ package }>"
+    user = "<{ no-infra-compute-user }>"
+  }
+}